+++ src/hdmf/spec/spec.py
@@ -41,7 +41,7 @@ class DtypeHelper:
         'object': ['object'],
         'region': ['region'],
         'numeric': ['numeric'],
-        'isodatetime': ["isodatetime", "datetime"]
+        'isodatetime': ["isodatetime", "datetime", "date"]
     }
 
     # List of recommended primary dtype strings. These are the keys of primary_dtype_string_synonyms